gendb/requirements.txt
flask
flask-wtf
flask-sqlalchemy
flask-migrate
flask-login
flask-bootstrap
mysqlclient